+++ html/projet.html
@@ -74,7 +74,7 @@
     <main class="main">
         <div class="shapes shape-1"></div>
         <div class="shapes shape-2 active-shape"></div>
-        <div class="section">
+        <section class="section">
 
             <div class="card big-card">
                 <a class="big-card__site" target="_blank" rel="noopener noreferrer">
@@ -110,8 +110,7 @@
 
                         <template id="tmp-card-skill">
                             <div class="card skill__card">
-                                <img class="skill__img" src="../assets/icon/html5.webp" alt="icon skill" width="80"
-                                    height="80">
+                                <img class="skill__img" width="80" height="80">
                                 <span class="skill__txt">Html 5</span>
                             </div>
                         </template>
@@ -119,7 +118,7 @@
                     </div>
                 </div>
             </div>
-        </div>
+        </section>
     </main>
     <footer class="footer">
         <p class="footer__p">
