+++ index.html
@@ -154,7 +154,7 @@
                 <template id="skill__card">
                     <div class="card skill__card">
                         <img class="skill__img" width="80" height="80">
-                        <span class="skill__txt">Html 5</span>
+                        <span class="skill__txt"></span>
                     </div>
                 </template>
                 <!-- Template -->
@@ -168,9 +168,9 @@
             <a href="./html/mentions_legales.html">Mentions légales</a>
         </p>
     </footer>
-    <!-- Effect multiText -->
+    <!-- MultiText Effect -->
     <script src="https://unpkg.com/typed.js@2.0.15/dist/typed.umd.js"></script>
-    <!-- Effect parralax -->
+    <!-- Parralax Effect -->
     <script src="./js/vanilla-tilt.min.js"></script>
 </body>
