+++ js/app.js
@@ -1,9 +1,11 @@
 import { fetchData } from "./fetch/fetchData.js"
+import { headerEffect, sliderNav, themeMode } from "./functions/functions.js"
 
 const projets = await fetchData("../DB/data.json")
+const skills = await fetchData("../DB/skills.json")
 
+/** Create cards */
 const cards = document.querySelector("#cards")
-
 for (const projet of projets) {
     const cardClone = document.querySelector("#tmp-card").content.firstElementChild.cloneNode(true)
     cardClone.href = `./html/projet.html?id=${projet.id}`
@@ -20,9 +22,7 @@ for (const projet of projets) {
 
     cards.append(cardClone)
 }
-
-const skills = await fetchData("../DB/skills.json")
-
+/** Create cards skill */
 const skillContent = document.querySelector("#skill__content")
 for (const skill of skills) {
     const skillClone = document.querySelector("#skill__card").content.firstElementChild.cloneNode(true)
@@ -34,32 +34,14 @@ for (const skill of skills) {
     skillContent.append(skillClone)
 }
 
-// Effect nav-slider ---------------------------------------------------------------------------------------------
-const btnToggle = document.querySelector(".btn-toggle")
-const navSlider = document.querySelector(".nav-slider")
-const navSliderLink = document.querySelectorAll(".nav-slider__link")
+/** ThemeMode  */
+themeMode()
+/** Nav-slider Effect */
+sliderNav()
+/** Header Effect */
+headerEffect()
 
-btnToggle.addEventListener("click", toggleNav)
-navSliderLink.forEach(a => a.addEventListener("click", () => toggleNav()))
-
-function toggleNav() {
-    btnToggle.classList.toggle("active")
-    navSlider.classList.toggle("active")
-}
-
-// Effect header ---------------------------------------------------------------------------------------------
-const header = document.querySelector('.header');
-
-window.addEventListener('scroll', () => {
-    // console.log(window.scrollY);
-    if (window.scrollY > 20) {
-        header.classList.add("active")
-    } else {
-        header.classList.remove("active")
-    }
-})
-
-// Effect nav ---------------------------------------------------------------------------------------------
+/** Nav Effect */
 const sections = document.querySelectorAll('.section')
 const links = document.querySelectorAll('.nav__link')
 
@@ -87,15 +69,14 @@ sections.forEach(section => {
     observer.observe(section)
 })
 
-
-// Active all shape > 1000px ---------------------------------------------------------------------------------------------
+/** Active all shape > 1000px  */
 if (window.innerWidth > 1000) {
     let shape = document.querySelectorAll(".shape-2")
     // console.log(shape);
     shape.forEach((a) => a.classList.add('active-shape'))
 }
 
-// Effect slide card ---------------------------------------------------------------------------------------------
+/** Effect slide card */
 const ratio = 0
 const options1 = {
     root: null,
@@ -121,50 +102,7 @@ document.querySelectorAll('.reveal-up').forEach((r) => {
     observer1.observe(r)
 })
 
-// DarkMode -------------------------------------------------------------------------------------------------------
-const theme = document.querySelector('.theme')
-const body = document.querySelector('body')
-const h3 = document.querySelectorAll('h3')
-const card = document.querySelectorAll('.card')
-const sun = document.querySelector('.sun')
-const moon = document.querySelector('.moon')
-
-const themeStorage = localStorage.getItem('darkMode')
-if (themeStorage === 'dark') {
-    darkMode()
-} else {
-    lightMode()
-}
-
-theme.addEventListener('click', () => {
-    const themeStorage = localStorage.getItem('darkMode')
-    if (themeStorage === 'dark') {
-        lightMode()
-    } else {
-        darkMode()
-    }
-})
-
-function darkMode() {
-    localStorage.setItem('darkMode', 'dark')
-    moon.setAttribute('hidden', '')
-    sun.removeAttribute('hidden')
-    body.classList.add('dark')
-    card.forEach(c => c.classList.add('dark'))
-    h3.forEach(h => h.classList.add('dark'))
-    header.classList.add('dark')
-}
-
-function lightMode() {
-    localStorage.setItem('darkMode', 'light')
-    sun.setAttribute('hidden', '')
-    moon.removeAttribute('hidden')
-    body.classList.remove('dark')
-    card.forEach(c => c.classList.remove('dark'))
-    h3.forEach(h => h.classList.remove('dark'))
-    header.classList.remove('dark')
-}
-
+/** Multi Text */
 let typingEffect = new Typed(".multiText", {
     strings: ["Jimmy Clairy", "Développeur"],
     loop: true,
@@ -172,7 +110,8 @@ let typingEffect = new Typed(".multiText", {
     backSpeed: 80,
     backDelay: 1500
 })
-// Active parralax > 1000px ---------------------------------------------------------------------------------------------
+
+/** Active parralax > 1000px */
 if (window.innerWidth > 1000) {
     VanillaTilt.init(document.querySelectorAll(".parralax"), {
         max: 15,
