+++ js/projet.js
@@ -1,10 +1,13 @@
 import { fetchData } from "./fetch/fetchData.js"
+import { headerEffect, sliderNav, themeMode } from "./functions/functions.js"
 
+/** Recup id in url */
 const id = new URL(window.location).searchParams.get("id")
 
 const projets = await fetchData("../DB/data.json")
 const projet = projets.find(p => p.id === Number(id))
 
+/** Create card */
 const titleHead = document.querySelector("title")
 titleHead.innerText = `Jimmy Clairy | ${projet.name}`
 
@@ -43,87 +46,26 @@ if (projet.imgLighthouse) {
     skill.style.width = "100%"
 }
 
+/** Create skill cards */
 const skillContent = document.querySelector("#skill__content")
-
 for (const icon of projet.icon) {
     const cardClone = document.querySelector("#tmp-card-skill").content.firstElementChild.cloneNode(true)
     const skillImg = cardClone.querySelector(".skill__img")
     skillImg.src = `../assets/icon/${icon.img}`
+    skillImg.alt = `Icon ${icon.name}`
     const skillTxt = cardClone.querySelector(".skill__txt")
     skillTxt.innerText = icon.name
     skillContent.append(cardClone)
 }
 
-// Effect nav-slider ---------------------------------------------------------------------------------------------
-const btnToggle = document.querySelector(".btn-toggle")
-const navSlider = document.querySelector(".nav-slider")
-const navSliderLink = document.querySelectorAll(".nav-slider__link")
+/** ThemeMode */
+themeMode()
+/** Nav-slider Effect */
+sliderNav()
+/** Header Effect */
+headerEffect()
 
-btnToggle.addEventListener("click", toggleNav)
-navSliderLink.forEach(a => a.addEventListener("click", () => toggleNav()))
-
-function toggleNav() {
-    btnToggle.classList.toggle("active")
-    navSlider.classList.toggle("active")
-}
-
-// Effect header ---------------------------------------------------------------------------------------------
-const header = document.querySelector('.header');
-
-window.addEventListener('scroll', () => {
-    // console.log(window.scrollY);
-    if (window.scrollY > 20) {
-        header.classList.add("active")
-    } else {
-        header.classList.remove("active")
-    }
-})
-
-// DarkMode -------------------------------------------------------------------------------------------------------
-const theme = document.querySelector('.theme')
-const body = document.querySelector('body')
-const h3 = document.querySelectorAll('h3')
-const card = document.querySelectorAll('.card')
-const sun = document.querySelector('.sun')
-const moon = document.querySelector('.moon')
-
-const themeStorage = localStorage.getItem('darkMode')
-if (themeStorage === 'dark') {
-    darkMode()
-} else {
-    lightMode()
-}
-
-theme.addEventListener('click', () => {
-    const themeStorage = localStorage.getItem('darkMode')
-    if (themeStorage === 'dark') {
-        lightMode()
-    } else {
-        darkMode()
-    }
-})
-
-function darkMode() {
-    localStorage.setItem('darkMode', 'dark')
-    moon.setAttribute('hidden', '')
-    sun.removeAttribute('hidden')
-    body.classList.add('dark')
-    card.forEach(c => c.classList.add('dark'))
-    h3.forEach(h => h.classList.add('dark'))
-    header.classList.add('dark')
-}
-
-function lightMode() {
-    localStorage.setItem('darkMode', 'light')
-    sun.setAttribute('hidden', '')
-    moon.removeAttribute('hidden')
-    body.classList.remove('dark')
-    card.forEach(c => c.classList.remove('dark'))
-    h3.forEach(h => h.classList.remove('dark'))
-    header.classList.remove('dark')
-}
-
-// Effect slide ---------------------------------------------------------------------------------------------
+/** Slide Effect */
 const ratio = 0
 const options1 = {
     root: null,
