+++ sass/main.scss
@@ -17,6 +17,6 @@
 @import "pages/accueil/projets";
 @import "pages/accueil/skills";
 
-@import "pages/others/big-card";
+@import "pages/projet";
 
 @import "bases/animated";
